+++ filter/CsrfFilter.js
@@ -15,9 +15,7 @@ module.exports = class CsrfFilter extends Base {
     }
 
     isActive (action) {
-        return action.controller.user.auth.csrf
-            && action.isPost()
-            && super.isActive(action);
+        return action.user.auth.csrf && action.isPost() && super.isActive(action);
     }
 
     beforeAction (action) {